Added a basic homepage with a test video.
This video is loaded from your computer directly ! It uses the
ReactPlayer suggested by @not_pickle_rick and this seems to work just
fine (at least for now).
I am not that familiar with React best practices for now, so please edit
any of the code that I got wrong.